consolidate utils/error modules, add hls + http config helper to gst plugin config struct
