+++ nats/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod cloud_commands;
 pub mod cloud_publisher;
 pub mod cloud_worker;
-pub mod error;
 pub mod message;
 pub mod subjects;
 pub mod subscriber;
